nixbot_cli: keep terminal formatting helpers in one module
status wording, colors and durations move to term.py and the log URL
onto the API client, removing the duplicate status sets and the lazy
import from main that worked around the module cycle.
